+++ src/shared/src/logistics/redux/index.js
@@ -8,8 +8,6 @@ export {
   makeSelectTaskListItemsByUsername,
   selectAllTasks,
   selectAssignedTasks,
-  selectFilteredTaskLists,
-  selectFilteredUnassignedTasksNotCancelled,
   selectSelectedDate,
   selectTaskLists,
   selectTasksWithColor,
